minor edit
git-svn-id: https://www.nick-ux.org/diagham/svn/trunk@1573 b9410c57-6ef9-0310-98bc-a73ba792996a

+++ FQHE/src/HilbertSpace/BosonOnLattice.cc
@@ -341,7 +341,7 @@ unsigned long BosonOnLattice::Ad (unsigned long state, int q, double& coefficien
   else
     {
       ++this->TemporaryState[q];
-      coefficient = sqrt(this->TemporaryState[q]);
+      coefficient = sqrt((double)this->TemporaryState[q]);
     }
   return this->BosonToFermion(this->TemporaryState, this->TemporaryStateHighestBit);
 }
